Created extended-teleport
extended-teleport is the unified version of the former gaze-teleport, mouse-teleport, gamepad-teleport and universal-teleport, all in one component.
