+++ index.html
@@ -26,8 +26,7 @@
         <a-entity id="camera"
                   camera
                   position="0 1.6 0"
-                  gamepad-teleport
-                  universal-teleport="curveShootingSpeed: 20; cameraRig: #rig; teleportOrigin: [camera]"
+                  extended-teleport="curveShootingSpeed: 20; cameraRig: #rig; teleportOrigin: [camera]"
                   look-controls="pointerLockEnabled: true;">
             <a-entity cursor
                       position="0 0 -1"
