+++ src/js/index.js
@@ -3,7 +3,4 @@ import 'aframe-extras';
 import 'aframe-teleport-controls';
 import 'aframe-environment-component';
 
-import { gazeTeleport } from './components/gaze-teleport';
-import { mouseTeleport } from './components/mouse-teleport';
-import { gamepadTeleport } from './components/gamepad-teleport';
-import { universalTeleport } from './components/universal-teleport';
+import { extendedTeleport} from './components/extended-teleport';
